fix(npc): SayImage conversation image-count is a byte, not int
CScriptMan::OnSayImage@0x6dc310 (case 1 of CScriptMan::OnScriptMessage)
reads the image count via CInPacket::Decode1 (1 byte) at 0x6dc3d9 before
looping DecodeStr per image. Atlas wrote the count as WriteInt (4 bytes),
desyncing the client's per-image string reads. Switch to WriteByte.

+++ libs/atlas-packet/npc/clientbound/conversation.go
@@ -116,7 +116,7 @@ type SayImageConversationDetail struct {
 func (s *SayImageConversationDetail) Encode(l logrus.FieldLogger, _ context.Context) func(options map[string]interface{}) []byte {
 	w := response.NewWriter(l)
 	return func(options map[string]interface{}) []byte {
-		w.WriteInt(uint32(len(s.Images)))
+		w.WriteByte(byte(len(s.Images)))
 		for _, image := range s.Images {
 			w.WriteAsciiString(image)
 		}
